+++ app/src/test/java/com/techcourse/AnnotationHandlerAdapterTest.java
@@ -2,6 +2,8 @@ package com.techcourse;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
 import org.junit.jupiter.api.Test;
