+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/AnnotationHandlerAdapter.java
@@ -1,7 +1,6 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.tobe;
 
 import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
